django/project/O2_pjt_accounts/requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2022.5
sqlparse==0.4.3
